refactor(runtime): isolate shell command structures
