+++ src/codex_plugin_scanner/guard/runtime/command_tokens.py
@@ -1,0 +1,103 @@
+"""Side-effect-free shell token and transparent wrapper parsing."""
+
+from __future__ import annotations
+
+import re
+import shlex
+
+_ENV_ASSIGNMENT_PATTERN = re.compile(r"^(?P<name>[A-Za-z_][A-Za-z0-9_]*)=.*$", re.DOTALL)
+_SUDO_OPTIONS_WITH_VALUES = frozenset({"-C", "-D", "-g", "-h", "-p", "-R", "-r", "-T", "-t", "-u"})
+_SUDO_LONG_OPTIONS_WITH_VALUES = frozenset(
+    {
+        "--chdir",
+        "--chroot",
+        "--close-from",
+        "--command-timeout",
+        "--group",
+        "--host",
+        "--login-class",
+        "--prompt",
+        "--role",
+        "--type",
+        "--user",
+    }
+)
+
+
+def executable_name(value: str | None) -> str | None:
+    """Return a normalized executable basename."""
+
+    if value is None:
+        return None
+    return value.replace("\\", "/").rsplit("/", 1)[-1].lower()
+
+
+def shell_tokens(command: str) -> tuple[tuple[str, ...], bool]:
+    """Tokenize shell text, reporting whether strict parsing succeeded."""
+
+    try:
+        return tuple(shlex.split(command, posix=True, comments=False)), True
+    except ValueError:
+        return tuple(command.split()), False
+
+
+def leading_environment(tokens: tuple[str, ...]) -> tuple[tuple[str, ...], int, tuple[str, ...]]:
+    """Return leading environment names, executable index, and wrappers."""
+
+    names: list[str] = []
+    wrappers: list[str] = []
+    index = 0
+    while index < len(tokens):
+        match = _ENV_ASSIGNMENT_PATTERN.fullmatch(tokens[index])
+        while match is not None:
+            names.append(match.group("name"))
+            index += 1
+            if index >= len(tokens):
+                return tuple(names), index, tuple(wrappers)
+            match = _ENV_ASSIGNMENT_PATTERN.fullmatch(tokens[index])
+        executable = executable_name(tokens[index])
+        if executable == "env":
+            wrappers.append("env")
+            index = _after_env_options(tokens, index + 1)
+            continue
+        if executable == "sudo":
+            wrappers.append("sudo")
+            index = _after_sudo_options(tokens, index + 1)
+            continue
+        break
+    return tuple(names), index, tuple(wrappers)
+
+
+def _after_env_options(tokens: tuple[str, ...], index: int) -> int:
+    while index < len(tokens):
+        token = tokens[index]
+        if token == "--":
+            return index + 1
+        if token in {"-u", "-C", "--unset", "--chdir"}:
+            index = min(index + 2, len(tokens))
+            continue
+        if token in {"-i", "-0", "--ignore-environment", "--null"} or token.startswith(("--unset=", "--chdir=")):
+            index += 1
+            continue
+        if _ENV_ASSIGNMENT_PATTERN.fullmatch(token) is not None:
+            return index
+        return index
+    return index
+
+
+def _after_sudo_options(tokens: tuple[str, ...], index: int) -> int:
+    while index < len(tokens):
+        token = tokens[index]
+        if token == "--":
+            return index + 1
+        option_name = token.split("=", 1)[0]
+        if option_name in _SUDO_OPTIONS_WITH_VALUES or option_name in _SUDO_LONG_OPTIONS_WITH_VALUES:
+            index += 1 if "=" in token else 2
+            continue
+        if token.startswith("-"):
+            index += 1
+            continue
+        if _ENV_ASSIGNMENT_PATTERN.fullmatch(token) is not None:
+            return index
+        return index
+    return index
